components/data-preprocessor/requirements.txt
pandas==1.3.5
scikit-learn==1.0.2
google-cloud-storage==2.2.0